[#361] Extract raw data to other file
