+++ frontend/src/modules/landing/new-home.jsx
@@ -26,537 +26,14 @@ import "./new-styles.scss";
 import "react-multi-carousel/lib/styles.css";
 import moment from "moment";
 import imageNotFound from "../../images/image-not-found.png";
-import TrimText from "../../utils/trim";
+import { TrimText } from "../../utils/string";
+import {
+  popularTopics,
+  featuredContents,
+  ourCommunity,
+  benefit,
+} from "./new-home-static-content";
 
-const popularTopics = [
-  {
-    id: 1,
-    topic: "Product by Design",
-    count: 65,
-    summary: [
-      {
-        count: 12,
-        type: "policy",
-      },
-      {
-        count: 13,
-        type: "events",
-      },
-      {
-        count: 14,
-        type: "technical resources",
-      },
-      {
-        count: 14,
-        type: "technology",
-      },
-      {
-        count: 12,
-        type: "action plan",
-      },
-      {
-        count: 14,
-        type: "initiatives",
-      },
-      {
-        count: 14,
-        type: "financial resources",
-      },
-    ],
-    items: [
-      {
-        id: 1,
-        type: "policy",
-        title: "Paullum deliquit, ponderibus modulisque suis ratio utitur.",
-        description:
-          "Integer legentibus erat a ante historiarum dapibus. Idque Caesaris facere voluntate liceret: sese habere. Ambitioni dedisse scripsisse iudicaretur.",
-      },
-      {
-        id: 2,
-        type: "technology",
-        title: "Cum sociis natoque penatibus et magnis dis parturient.",
-        description: "At nos hinc posthac, sitientis piros Afros.",
-      },
-      {
-        id: 3,
-        type: "event",
-        title: "Nihilne te nocturnum praesidium Palati, nihil urbis vigiliae.",
-        description:
-          "Ambitioni dedisse scripsisse iudicaretur. Quis aute iure reprehenderit in voluptate velit esse. Magna pars studiorum, prodita quaerimus. ",
-      },
-    ],
-  },
-  {
-    id: 2,
-    topic: "Waste Management",
-    count: 61,
-    summary: [
-      {
-        count: 12,
-        type: "policy",
-      },
-      {
-        count: 13,
-        type: "events",
-      },
-      {
-        count: 12,
-        type: "technical resources",
-      },
-      {
-        count: 12,
-        type: "technology",
-      },
-      {
-        count: 12,
-        type: "action plan",
-      },
-      {
-        count: 12,
-        type: "initiatives",
-      },
-      {
-        count: 12,
-        type: "financial resources",
-      },
-    ],
-    items: [
-      {
-        id: 1,
-        type: "policy",
-        title: "Paullum deliquit, ponderibus modulisque suis ratio utitur.",
-        description:
-          "Integer legentibus erat a ante historiarum dapibus. Idque Caesaris facere voluntate liceret: sese habere. Ambitioni dedisse scripsisse iudicaretur.",
-      },
-      {
-        id: 2,
-        type: "technology",
-        title: "Cum sociis natoque penatibus et magnis dis parturient.",
-        description: "At nos hinc posthac, sitientis piros Afros.",
-      },
-      {
-        id: 3,
-        type: "event",
-        title: "Nihilne te nocturnum praesidium Palati, nihil urbis vigiliae.",
-        description:
-          "Ambitioni dedisse scripsisse iudicaretur. Quis aute iure reprehenderit in voluptate velit esse. Magna pars studiorum, prodita quaerimus. ",
-      },
-    ],
-  },
-  {
-    id: 3,
-    topic: "Plastics",
-    count: 57,
-    summary: [
-      {
-        count: 12,
-        type: "policy",
-      },
-      {
-        count: 13,
-        type: "events",
-      },
-      {
-        count: 14,
-        type: "technical resources",
-      },
-      {
-        count: 14,
-        type: "technology",
-      },
-      {
-        count: 12,
-        type: "action plan",
-      },
-      {
-        count: 14,
-        type: "initiatives",
-      },
-      {
-        count: 14,
-        type: "financial resources",
-      },
-    ],
-    items: [
-      {
-        id: 1,
-        type: "policy",
-        title: "Paullum deliquit, ponderibus modulisque suis ratio utitur.",
-        description:
-          "Integer legentibus erat a ante historiarum dapibus. Idque Caesaris facere voluntate liceret: sese habere. Ambitioni dedisse scripsisse iudicaretur.",
-      },
-      {
-        id: 2,
-        type: "technology",
-        title: "Cum sociis natoque penatibus et magnis dis parturient.",
-        description: "At nos hinc posthac, sitientis piros Afros.",
-      },
-      {
-        id: 3,
-        type: "event",
-        title: "Nihilne te nocturnum praesidium Palati, nihil urbis vigiliae.",
-        description:
-          "Ambitioni dedisse scripsisse iudicaretur. Quis aute iure reprehenderit in voluptate velit esse. Magna pars studiorum, prodita quaerimus. ",
-      },
-    ],
-  },
-  {
-    id: 4,
-    topic: "Marine Litter",
-    count: 56,
-    summary: [
-      {
-        count: 12,
-        type: "policy",
-      },
-      {
-        count: 13,
-        type: "events",
-      },
-      {
-        count: 14,
-        type: "technical resources",
-      },
-      {
-        count: 14,
-        type: "technology",
-      },
-      {
-        count: 12,
-        type: "action plan",
-      },
-      {
-        count: 14,
-        type: "initiatives",
-      },
-      {
-        count: 14,
-        type: "financial resources",
-      },
-    ],
-    items: [
-      {
-        id: 1,
-        type: "policy",
-        title: "Paullum deliquit, ponderibus modulisque suis ratio utitur.",
-        description:
-          "Integer legentibus erat a ante historiarum dapibus. Idque Caesaris facere voluntate liceret: sese habere. Ambitioni dedisse scripsisse iudicaretur.",
-      },
-      {
-        id: 2,
-        type: "technology",
-        title: "Cum sociis natoque penatibus et magnis dis parturient.",
-        description: "At nos hinc posthac, sitientis piros Afros.",
-      },
-      {
-        id: 3,
-        type: "event",
-        title: "Nihilne te nocturnum praesidium Palati, nihil urbis vigiliae.",
-        description:
-          "Ambitioni dedisse scripsisse iudicaretur. Quis aute iure reprehenderit in voluptate velit esse. Magna pars studiorum, prodita quaerimus. ",
-      },
-    ],
-  },
-  {
-    id: 5,
-    topic: "Capacity Building",
-    count: 50,
-    summary: [
-      {
-        count: 12,
-        type: "policy",
-      },
-      {
-        count: 13,
-        type: "events",
-      },
-      {
-        count: 14,
-        type: "technical resources",
-      },
-      {
-        count: 14,
-        type: "technology",
-      },
-      {
-        count: 12,
-        type: "action plan",
-      },
-      {
-        count: 14,
-        type: "initiatives",
-      },
-      {
-        count: 14,
-        type: "financial resources",
-      },
-    ],
-    items: [
-      {
-        id: 1,
-        type: "policy",
-        title: "Paullum deliquit, ponderibus modulisque suis ratio utitur.",
-        description:
-          "Integer legentibus erat a ante historiarum dapibus. Idque Caesaris facere voluntate liceret: sese habere. Ambitioni dedisse scripsisse iudicaretur.",
-      },
-      {
-        id: 2,
-        type: "technology",
-        title: "Cum sociis natoque penatibus et magnis dis parturient.",
-        description: "At nos hinc posthac, sitientis piros Afros.",
-      },
-      {
-        id: 3,
-        type: "event",
-        title: "Nihilne te nocturnum praesidium Palati, nihil urbis vigiliae.",
-        description:
-          "Ambitioni dedisse scripsisse iudicaretur. Quis aute iure reprehenderit in voluptate velit esse. Magna pars studiorum, prodita quaerimus. ",
-      },
-    ],
-  },
-  {
-    id: 6,
-    topic: "Source to Sea",
-    count: 48,
-    summary: [
-      {
-        count: 12,
-        type: "policy",
-      },
-      {
-        count: 13,
-        type: "events",
-      },
-      {
-        count: 14,
-        type: "technical resources",
-      },
-      {
-        count: 14,
-        type: "technology",
-      },
-      {
-        count: 12,
-        type: "action plan",
-      },
-      {
-        count: 14,
-        type: "initiatives",
-      },
-      {
-        count: 14,
-        type: "financial resources",
-      },
-    ],
-    items: [
-      {
-        id: 1,
-        type: "policy",
-        title: "Paullum deliquit, ponderibus modulisque suis ratio utitur.",
-        description:
-          "Integer legentibus erat a ante historiarum dapibus. Idque Caesaris facere voluntate liceret: sese habere. Ambitioni dedisse scripsisse iudicaretur.",
-      },
-      {
-        id: 2,
-        type: "technology",
-        title: "Cum sociis natoque penatibus et magnis dis parturient.",
-        description: "At nos hinc posthac, sitientis piros Afros.",
-      },
-      {
-        id: 3,
-        type: "event",
-        title: "Nihilne te nocturnum praesidium Palati, nihil urbis vigiliae.",
-        description:
-          "Ambitioni dedisse scripsisse iudicaretur. Quis aute iure reprehenderit in voluptate velit esse. Magna pars studiorum, prodita quaerimus. ",
-      },
-    ],
-  },
-];
-const featuredContents = [
-  {
-    image: null,
-    type: "data hub",
-    title: "Morbi odio eros, volutpat ut pharetra vitae, lobortis sed nibh.",
-    description:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec tempor ante ac leo cursus, quis fringilla elit sagittis. Maecenas ac maximus massa...",
-    bookmark: [
-      {
-        image: "",
-        name: "Name 1",
-      },
-      {
-        image: "",
-        name: "Name 2",
-      },
-      {
-        image: "",
-        name: "Name 3",
-      },
-      {
-        image: "",
-        name: "Name 4",
-      },
-    ],
-  },
-  {
-    image: null,
-    type: "policy",
-    title: "Nihilne te nocturnum praesidium Palati, nihil urbis vigiliae.",
-    description:
-      "Integer legentibus erat a ante historiarum dapibus. Idque Caesaris facere voluntate liceret: sese habere. Ambitioni dedisse scripsisse iudicaretur. Fabio vel iudice vincam,",
-    bookmark: [
-      {
-        image: "",
-        name: "Name 1",
-      },
-      {
-        image: "",
-        name: "Name 2",
-      },
-      {
-        image: "",
-        name: "Name 3",
-      },
-      {
-        image: "",
-        name: "Name 4",
-      },
-    ],
-  },
-  {
-    image: "fc-initiative.png",
-    type: "initiative",
-    title: "Morbi odio eros, volutpat ut pharetra vitae, lobortis sed nibh.",
-    description:
-      "Inmensae subtilitatis, obscuris et malesuada fames. Ut enim ad minim veniam, quis nostrud exercitation. Quisque ut dolor gravida, placerat libero vel, euismod. Salutantibus vitae elit libero, a pharetra augue. Vivamus sagittis lacus vel augue laoreet rutrum faucibus. Quam temere in vitiis, legem sancimus haerentia. Fabio vel iudice vincam, sunt in culpa qui officia. Quis aute iure reprehenderit in voluptate velit esse. Non equidem invideo, miror magis posuere velit aliquet. Quisque placerat facilisis egestas cillum dolore. Praeterea iter est quasdam res quas ex communi.",
-    bookmark: [
-      {
-        image: "",
-        name: "Name 1",
-      },
-      {
-        image: "",
-        name: "Name 2",
-      },
-      {
-        image: "",
-        name: "Name 3",
-      },
-      {
-        image: "",
-        name: "Name 4",
-      },
-      {
-        image: "",
-        name: "Name 4",
-      },
-    ],
-  },
-  {
-    image: null,
-    type: "technical resource",
-    title: "Contra legem facit qui id facit quod lex prohibet.",
-    description:
-      "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec tempor ante ac leo cursus, quis fringilla elit sagittis. Maecenas ac maximus massa...",
-    bookmark: [
-      {
-        image: "",
-        name: "Name 1",
-      },
-      {
-        image: "",
-        name: "Name 2",
-      },
-      {
-        image: "",
-        name: "Name 3",
-      },
-      {
-        image: "",
-        name: "Name 4",
-      },
-    ],
-  },
-  {
-    image: null,
-    type: "story map",
-    title: "At nos hinc posthac, sitientis piros Afros.",
-    description:
-      "Tu quoque, Brute, fili mi, nihil timor populi, nihil! Quisque placerat facilisis egestas cillum dolore. Hi omnes lingua, institutis, legibus inter se differunt. Paullum deliquit, ponderibus",
-    bookmark: [
-      {
-        image: "",
-        name: "Name 1",
-      },
-      {
-        image: "",
-        name: "Name 2",
-      },
-      {
-        image: "",
-        name: "Name 3",
-      },
-      {
-        image: "",
-        name: "Name 4",
-      },
-    ],
-  },
-];
-const ourCommunity = [
-  {
-    type: "stakeholder",
-    about:
-      "Sed haec quis possit intrepidus aestimare tellus. Cum sociis natoque penatibus et magnis dis parturient.",
-    name: "John Malkovich",
-    role: "Entity Role -  Entity",
-  },
-  {
-    type: "stakeholder",
-    about: "Salutantibus vitae elit libero, a pharetra augue.",
-    name: "Bertrand Lacaze",
-    role: "Entity Role -  Entity",
-  },
-  {
-    type: "stakeholder",
-    about:
-      "Contra legem facit qui id facit quod lex prohibet. Quid securi etiam tamquam eu fugiat nulla pariatur.",
-    name: "Xavier Mendoza",
-    role: "Entity Role -  Entity",
-  },
-  {
-    type: "entity",
-    about:
-      "Contra legem facit qui id facit quod lex prohibet. Quid securi etiam tamquam eu fugiat nulla pariatur.",
-    name: "Black Forest Solutions",
-    role: null,
-  },
-  {
-    type: "stakeholder",
-    about:
-      "Contra legem facit qui id facit quod lex prohibet. Quid securi etiam tamquam eu fugiat nulla pariatur.",
-    name: "Xavier",
-    role: "Entity Role -  Entity",
-  },
-];
-const benefit = [
-  {
-    title: "GPML Partnership",
-    childs: [
-      "Tap into a global network of like-minded members​",
-      "Discover opportunities to showcase your work​",
-      "Avoid duplication of effort and optimise impact​",
-    ],
-  },
-  {
-    title: "GPML Digital Platform​",
-    childs: [
-      "Access a data hub to guide efforts towards SDGs and more",
-      "Utilise an array of resources at your fingertips​",
-      "Network with other stakeholders​",
-    ],
-  },
-];
 const cardSvg = [
   {
     color: "#4DA687",
@@ -623,6 +100,7 @@ const cardSvg = [
     ),
   },
 ];
+
 const responsive = {
   superLargeDesktop: {
     breakpoint: { max: 4000, min: 1200 },
